refactor: search persistence after new incomming data
